D2: change IPool and derivatives to work with size_t

+++ src/ocean/io/select/client/Scheduler.d
@@ -144,7 +144,7 @@ public class Scheduler ( EventData ) : TimerEventTimeoutManager
 
         ***********************************************************************/
 
-        public uint object_pool_index;
+        public size_t object_pool_index;
 
 
         /***********************************************************************
